requirements.txt
rastervision==0.31.0
rastervision_pipeline==0.31.2
shapely==2.0.5
geopandas==1.0.1
numpy==1.26.4
pillow==10.4.0
pyproj==3.6.1
rasterio==1.3.10
pystac==1.10.1
scikit-learn==1.5.1
scipy==1.14.0
opencv-python-headless==4.10.0.84
tqdm==4.66.5
xarray==2024.7.0
scikit-image==0.24.0
boto3==1.34.155
stackstac==0.5.0
humanize==4.10.0
triangle==20230923
# gdal==3.9.2 #se hace con conda
gdown==5.2.0
utils==
azure.storage.blob == 
geopandas==1.0.1    

